Move shared test setup to lib/galaxy_test/driver/integration_setup.py
